Rename service loader classes for conversions
