+++ sqrl-engines/sqrl-engine-flink/src/main/java/com/datasqrl/engine/stream/flink/sql/FlinkRelToSqlNode.java
@@ -2,19 +2,19 @@ package com.datasqrl.engine.stream.flink.sql;
 
 import com.datasqrl.FlinkExecutablePlan.FlinkQuery;
 import com.datasqrl.calcite.Dialect;
-import com.datasqrl.calcite.convert.SqlConverter;
+import com.datasqrl.calcite.convert.RelToSqlNode;
 import com.google.auto.service.AutoService;
-import com.google.common.base.Preconditions;
+
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
+
 import lombok.Value;
 import org.apache.calcite.rel.RelNode;
 import org.apache.calcite.rel.rel2sql.FlinkRelToSqlConverter;
 import org.apache.calcite.sql.SqlNode;
 
-@AutoService(SqlConverter.class)
-public class FlinkSqlConverter implements SqlConverter {
+@AutoService(RelToSqlNode.class)
+public class FlinkRelToSqlNode implements RelToSqlNode {
 
   @Override
   public SqlNodes convert(RelNode relNode) {
