+++ sqrl-engines/sqrl-engine-flink/src/main/java/com/datasqrl/engine/stream/flink/sql/RelToFlinkSql.java
@@ -1,8 +1,7 @@
 package com.datasqrl.engine.stream.flink.sql;
 
-import com.datasqrl.calcite.convert.SqlConverter.SqlNodes;
-import com.datasqrl.calcite.convert.SqlToString.SqlStrings;
-import com.datasqrl.engine.ExecutionEngine;
+import com.datasqrl.calcite.convert.RelToSqlNode.SqlNodes;
+import com.datasqrl.calcite.convert.SqlNodeToString.SqlStrings;
 import com.datasqrl.engine.stream.flink.sql.calcite.FlinkDialect;
 import com.datasqrl.engine.stream.flink.sql.model.QueryPipelineItem;
 import java.util.function.UnaryOperator;
