+++ sqrl-planner/sqrl-common/src/main/java/com/datasqrl/calcite/convert/RelToSqlNode.java
@@ -5,7 +5,7 @@ import org.apache.calcite.rel.RelNode;
 import org.apache.calcite.sql.SqlNode;
 
 //Service loader interface
-public interface SqlConverter {
+public interface RelToSqlNode {
   SqlNodes convert(RelNode relNode);
 
   Dialect getDialect();
