+++ sqrl-planner/sqrl-common/src/main/java/com/datasqrl/calcite/convert/SqlConverterFactory.java
@@ -5,8 +5,8 @@ import com.datasqrl.util.ServiceLoaderDiscovery;
 
 public class SqlConverterFactory {
 
-  public static SqlConverter get(Dialect dialect) {
-    return ServiceLoaderDiscovery.get(SqlConverter.class, e->e.getDialect().name(),
+  public static RelToSqlNode get(Dialect dialect) {
+    return ServiceLoaderDiscovery.get(RelToSqlNode.class, e->e.getDialect().name(),
         dialect.name());
   }
 }
